fix(inference): require llama.cpp served aliases
